requirements.txt
fastapi
uvicorn
playwright
requests
aiohttp
aiodns
cchardet
ujson
python-jose
passlib
python-multipart
cryptography
bcrypt
python-dotenv
